pyproject.toml
[tool.poetry]
name = "gstk"
version = "0.0.0"
description = "Generative Story Toolkit"
authors = ["Justin McOmie <justin.mcomie@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gstk"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.2.1"
python-dotenv = "^1.0.0"
openai = {extras = ["datalib", "embeddings"], version = "^0.27.9"}
appdirs = "^1.4.4"
tiktoken = "^0.5.1"
langchain-experimental = "^0.0.43"
jsonschema = "^4.20.0"
faiss-cpu = "^1.7.4"
pre-commit = "^3.6.0"
jupyter = "^1.0.0"

[tool.poetry-dynamic-versioning]
enable = true

[tool.poetry.group.dev.dependencies]
setuptools = "^68.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120

[tool.isort]
profile = "black"
line_length = 120
extra_standard_library = ["pytest"]
